testing: factor out rest shared examples further
