+++ spec/helpers/unit/provider/elasticsearch_rest_shared_examples.rb
