+++ spec/unit/type/elasticsearch_pipeline_spec.rb
@@ -1,9 +1,9 @@
 require 'spec_helper'
 
-require_relative 'elasticsearch_rest_shared_examples'
+require_relative '../../helpers/unit/type/elasticsearch_rest_shared_examples'
 
 describe Puppet::Type.type(:elasticsearch_pipeline) do
   let(:resource_name) { 'test_pipeline' }
 
-  include_examples 'REST API types', 'pipeline'
+  include_examples 'REST API types', 'pipeline', :content
 end
